Add user presence. Simplify loading. Close #994
The user presence is important feature for all realtime social
applications. During the implementation I discovered, that transit-js
does not support undefined type. It converts it to the null. After the
initial disappointment I realised that the same behaviour as Firebase
has. So I had to reimplement loading HOC and realised again, that
loading HOC was too smart.

- Nice user presence added.
- Some Firebase Bolt rules were updated, look how easily we can make
new rules from yet existing. Now the list of app users is limited so it
can’t be scanned at whole.
- Add placeholder Bolt rule isFriend.
- Make loading pattern less smart and more explicit. Remove loading
higher order component.